# backend/docs/swagger.yaml
definitions:
...
  handler.APIResponse:
    properties:
      data: {}
      msg:
        example: ok
        type: string
    type: object
...
  handler.RequestResponse:
    properties:
      address:
        example: 北京市朝阳区幸福小区1号楼
        type: string
      appointment_time:
        type: string
      contact_name:
        example: 张大爷
        type: string
      contact_phone:
        example: "13800000008"
        type: string
      created_at:
        type: string
      description:
        example: 需要午餐送餐
        type: string
      feedback:
        type: string
      id:
        example: 1
        type: integer
      images:
        type: string
      rating:
        example: 5
        type: integer
      reject_reason:
        type: string
      request_no:
        example: REQ20260209001
        type: string
      service_type:
        example: meal
        type: string
      station_id:
        example: 1
        type: integer
      station_name:
        example: 朝阳区幸福街道服务站
        type: string
      status:
        example: dispatched
        type: string
      updated_at:
        type: string
      urgency:
        example: normal
        type: string
      user_id:
        example: 8
        type: integer
    type: object
...
  handler.rateRequest:
    properties:
      feedback:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
  handler.requestCreate:
    properties:
      address:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      images:
        items:
          type: string
        type: array
      lat:
        type: number
      lng:
        type: number
      request_no:
        type: string
      service_type:
        type: string
    required:
    - service_type
    type: object
...
  handler.updateStatusRequest:
    properties:
      reject_reason:
        type: string
      status:
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
paths:
...
  /b/requests:
    get:
      consumes:
      - application/json
      description: 获取当前用户的服务需求列表，支持按状态筛选和分页
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务需求列表
      tags:
      - c_request
      - b_request
  /b/requests/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取服务需求的详细信息
      parameters:
      - description: 需求ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: ID无效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 需求不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务需求详情
      tags:
      - c_request
      - b_request
  /b/requests/{id}/status:
    put:
      consumes:
      - application/json
      description: B端更新服务请求状态（如拒绝、完成等）
      parameters:
      - description: 需求ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新服务请求状态
      tags:
      - b_request
...
  /c/requests:
    get:
      consumes:
      - application/json
      description: 获取当前用户的服务需求列表，支持按状态筛选和分页
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务需求列表
      tags:
      - c_request
      - b_request
    post:
      consumes:
      - application/json
      description: C端用户创建新的服务需求，系统自动匹配站点并派发任务
      parameters:
      - description: 服务需求信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.requestCreate'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: 请求冲突
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 创建服务需求
      tags:
      - c_request
  /c/requests/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取服务需求的详细信息
      parameters:
      - description: 需求ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: ID无效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 需求不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务需求详情
      tags:
      - c_request
      - b_request
  /c/requests/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 取消待处理的服务需求
      parameters:
      - description: 需求ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: ID无效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: 状态不允许取消
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 取消服务需求
      tags:
      - c_request
  /c/requests/{id}/rate:
    post:
      consumes:
      - application/json
      description: 对已完成的服务进行评价
      parameters:
      - description: 需求ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评价信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.rateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 评价成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: 状态不允许评价
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 评价服务
      tags:
      - c_request
...
swagger: "2.0"